+++ src/ml_tooling/baseclass.py
@@ -14,10 +14,10 @@ from sklearn.pipeline import Pipeline
 from ml_tooling.config import config
 from ml_tooling.data.base_data import Dataset
 from ml_tooling.logging.logger import create_logger
-from ml_tooling.metrics.metric import Metrics
 from ml_tooling.result import ResultType
 from ml_tooling.result.result import Result
 from ml_tooling.result.result_group import ResultGroup
+from ml_tooling.search import BayesSearch
 from ml_tooling.search.base import Searcher
 from ml_tooling.search.gridsearch import GridSearch
 from ml_tooling.search.randomsearch import RandomSearch
@@ -132,7 +132,9 @@ class Model:
         return storage.get_list()
 
     @classmethod
-    def load_estimator(cls, storage: Storage, path: pathlib.Path) -> "Model":
+    def load_estimator(
+        cls, path: Union[str, pathlib.Path], storage: Storage = None
+    ) -> "Model":
         """
         Instantiates the class with a joblib pickled estimator.
 
@@ -140,7 +142,8 @@ class Model:
         ----------
         storage : Storage
             Storage class to load the estimator with
-        path: str, optional
+
+        path: str, pathlib.Path, optional
             Path to estimator pickle file
 
         Example
@@ -148,17 +151,24 @@ class Model:
         We can load a trained estimator from disk::
 
             storage = FileStorage('path/to/dir')
-            my_estimator = Model.load_estimator(storage, 'my_model.pkl')
+            my_estimator = Model.load_estimator('my_model.pkl', storage=storage)
 
         We now have a trained estimator loaded.
+
+        We can also use the default storage::
+
+            my_estimator = Model.load_estimator('my_model.pkl')
+
+        This will use the default FileStorage defined in Model.config.default_storage
 
         Returns
         -------
         Model
             Instance of Model with a saved estimator
         """
+        fs = config.default_storage if storage is None else storage
         filename = pathlib.Path(path).name
-        estimator = storage.load(filename)
+        estimator = fs.load(filename)
         instance = cls(estimator)
         logger.info(f"Loaded {instance.estimator_name}")
         return instance
@@ -306,6 +316,7 @@ class Model:
         cls,
         data: Dataset,
         estimators: Sequence[Estimator],
+        feature_pipeline: Pipeline = None,
         metrics: Union[str, List[str]] = "default",
         cv: Union[int, bool] = False,
         log_dir: str = None,
@@ -321,6 +332,9 @@ class Model:
 
         estimators: Sequence[Estimator]
             List of estimators to train
+
+        feature_pipeline: Pipeline
+            A pipeline for transforming features
 
         metrics: str, list of str
             Metric or list of metrics to use in scoring of estimators
@@ -339,7 +353,13 @@ class Model:
         List of Result objects
         """
 
-        results = _train_estimators(estimators, data, metrics, cv)
+        results = _train_estimators(
+            estimators=estimators,
+            feature_pipeline=feature_pipeline,
+            data=data,
+            metrics=metrics,
+            cv=cv,
+        )
 
         if log_dir:
             results.log(pathlib.Path(log_dir))
@@ -417,7 +437,6 @@ class Model:
         if isinstance(metrics, str):
             metrics = [self.default_metric] if metrics == "default" else [metrics]
 
-        score_metrics = Metrics.from_list(metrics)
         cv = cv if cv else None
 
         if cv:
@@ -441,7 +460,7 @@ class Model:
         self.result = Result.from_estimator(
             estimator=self.estimator,
             data=data,
-            metrics=score_metrics,
+            metrics=metrics,
             cv=cv,
             n_jobs=self.config.N_JOBS,
             verbose=self.config.VERBOSITY,
@@ -458,7 +477,8 @@ class Model:
         searcher: Searcher,
         metrics: Union[str, List[str]] = "default",
         cv: Optional[int] = None,
-    ) -> ResultGroup:
+        refit: bool = True,
+    ) -> Tuple["Model", ResultGroup]:
         """
         Runs a cross-validated search on the given estimators.
 
@@ -476,6 +496,9 @@ class Model:
 
         cv: int, optional
             Cross validation to use. Defaults to value in :attr:`config.CROSS_VALIDATION`
+
+        refit: bool
+            Whether or not to refit the best model
 
         Returns
         -------
@@ -497,6 +520,14 @@ class Model:
         self.result: ResultGroup = searcher.search(
             data, metrics, cv, n_jobs=config.N_JOBS, verbose=config.VERBOSITY
         )
+        best_estimator = Model(self.result.estimator)
+        logger.info(
+            f"Best estimator: {best_estimator.estimator_name} - "
+            f"{self.result.metrics.name}: {self.result.metrics.score}"
+        )
+
+        if refit:
+            best_estimator.score_estimator(data, metrics)
 
         logger.info("Done!")
 
@@ -504,7 +535,7 @@ class Model:
             result_file = self.result.log(self.config.RUN_DIR)
             logger.info(f"Saved run info at {result_file}")
 
-        return self.result
+        return best_estimator, self.result
 
     def gridsearch(
         self,
@@ -546,14 +577,9 @@ class Model:
 
         gs = GridSearch(self.estimator, param_grid=param_grid)
         logger.debug(f"Gridsearching using {param_grid}")
-        results = self._cross_validated_search(data, gs, metrics=metrics, cv=cv)
-
-        best_model = results[0].model
-
-        if refit:
-            best_model.score_estimator(data, metrics)
-
-        return best_model, results
+        return self._cross_validated_search(
+            data, gs, metrics=metrics, cv=cv, refit=refit
+        )
 
     def randomsearch(
         self,
@@ -603,16 +629,68 @@ class Model:
             self.estimator, param_grid=param_distributions, n_iter=n_iter
         )
 
-        results = self._cross_validated_search(
-            data=data, searcher=searcher, metrics=metrics, cv=cv
+        logger.debug("Random Searching using %s", param_distributions)
+
+        return self._cross_validated_search(
+            data=data, searcher=searcher, metrics=metrics, cv=cv, refit=refit
         )
 
-        best_model = Model(results[0].estimator)
+    def bayesiansearch(
+        self,
+        data: Dataset,
+        param_distributions: dict,
+        metrics: Union[str, List[str]] = "default",
+        cv: Optional[int] = None,
+        n_iter: int = 10,
+        refit: bool = True,
+    ) -> Tuple["Model", ResultGroup]:
+        """
+        Runs a cross-validated Bayesian Search on the estimator with a randomized
+        sampling of the passed parameter distributions
 
-        if refit:
-            best_model.score_estimator(data, metrics)
+        Parameters
+        ----------
+        data: Dataset
+            An instance of a DataSet object
 
-        return best_model, results
+        param_distributions: dict
+            Parameter distributions to use for randomizing search. Should be a dictionary
+            of param_names -> one of
+            - :class:`ml_tooling.search.Integer`
+            - :class:`ml_tooling.search.Categorical`
+            - :class:`ml_tooling.search.Real`
+
+        metrics: str, list of str
+            Metrics to use for scoring. "default" sets metric equal to
+            :attr:`self.default_metric`. First metric is used to sort results.
+
+        cv: int, optional
+            Cross validation to use. Defaults to value in :attr:`config.CROSS_VALIDATION`
+
+        n_iter: int
+            Number of parameter settings that are sampled.
+
+        refit: bool
+            Whether or not to refit the best model
+
+        Returns
+        -------
+        best_estimator: Model
+            Best estimator as found by the Bayesian Search
+
+        result_group: ResultGroup
+            ResultGroup object containing each individual score
+        """
+
+        searcher = BayesSearch(
+            self.estimator, param_grid=param_distributions, n_iter=n_iter
+        )
+
+        logger.debug("Bayesian Search using %s", param_distributions)
+
+        return self._cross_validated_search(
+            data=data, searcher=searcher, metrics=metrics, cv=cv, refit=refit
+        )
 
     @contextmanager
     def log(self, run_directory: str):
@@ -672,6 +750,7 @@ def _train_estimators(
     data: Dataset,
     metrics: Union[str, List[str]],
     cv: Any,
+    feature_pipeline: Pipeline = None,
 ) -> ResultGroup:
     """
     Sequentially train a series of models and create a ResultGroup of the results
@@ -697,7 +776,7 @@ def _train_estimators(
 
     results = []
     for i, estimator in enumerate(estimators, start=1):
-        challenger_estimator = Model(estimator)
+        challenger_estimator = Model(estimator, feature_pipeline=feature_pipeline)
         logger.info(
             f"Training estimator {i}/{len(estimators)}: {challenger_estimator.estimator_name}"
         )