+++ src/ml_tooling/transformers/__init__.py
@@ -1,5 +1,6 @@
 from .binner import Binner
-from .categorical import ToCategorical, Binarize
+from .to_categorical import ToCategorical
+from .binarize import Binarize
 from .date_encoder import DateEncoder
 from .feature_union import DFFeatureUnion
 from .fillna import FillNA
@@ -9,12 +10,13 @@ from .rare_feature import RareFeatureEncoder
 from .renamer import Renamer
 from .scaler import DFStandardScaler
 from .select import Select
+from .pipeline import Pipeline
 
 
 __all__ = [
+    "Binarize",
     "Binner",
     "ToCategorical",
-    "Binarize",
     "DateEncoder",
     "DFFeatureUnion",
     "FillNA",
@@ -25,4 +27,5 @@ __all__ = [
     "Renamer",
     "DFStandardScaler",
     "Select",
+    "Pipeline",
 ]