+++ tests/test_transformers.py
@@ -13,7 +13,6 @@ from ml_tooling.result import Result
 from ml_tooling.utils import TransformerError
 from ml_tooling.transformers import (
     Select,
-    FillNA,
     ToCategorical,
     FuncTransformer,
     Binner,
@@ -90,182 +89,14 @@ class TestDFSelector:
         result = model.score_estimator(train_iris_dataset)
         assert isinstance(result, Result)
 
-
-class TestFillNA:
-    @pytest.mark.parametrize("value, strategy", [(None, None), (0, "mean")])
-    def test_fillna_raises_error(self, numerical_na: pd.DataFrame, value, strategy):
-        with pytest.raises(TransformerError):
-            FillNA(value=value, strategy=strategy).fit(numerical_na)
-
-    def test_fillna_returns_dataframe_unchanged_if_no_nans(
-        self, categorical: pd.DataFrame
-    ):
-        imputer = FillNA("Unknown")
-        result = imputer.fit_transform(categorical)
-        pd.testing.assert_frame_equal(result, categorical)
-
-    @pytest.mark.parametrize(
-        "value, strategy, expected",
-        [
-            (
-                "Unknown",
-                None,
-                pd.DataFrame(
-                    {
-                        "category_a": ["a1", "Unknown", "a3", "a1"],
-                        "category_b": ["Unknown", "b2", "b3", "b1"],
-                    }
-                ),
-            ),
-            (
-                None,
-                "most_freq",
-                pd.DataFrame(
-                    {
-                        "category_a": ["a1", "a1", "a3", "a1"],
-                        "category_b": ["b1", "b2", "b3", "b1"],
-                    }
-                ),
-            ),
-        ],
-    )
-    def test_fillna_imputes_categorical_na_correct(
-        self, categorical_na: pd.DataFrame, value: Any, strategy: Any, expected: Any
-    ):
-        imputer = FillNA(value=value, strategy=strategy)
-        result = imputer.fit_transform(categorical_na)
-        pd.testing.assert_frame_equal(result, expected)
-
-    @pytest.mark.parametrize(
-        "value, strategy, expected",
-        [
-            (
-                None,
-                "mean",
-                pd.DataFrame(
-                    {"number_a": [3.0, 2.0, 3.0, 4.0], "number_b": [5.0, 6.0, 7.0, 6.0]}
-                ),
-            ),
-            (
-                None,
-                "median",
-                pd.DataFrame(
-                    {"number_a": [3.0, 2.0, 3.0, 4.0], "number_b": [5.0, 6.0, 7.0, 6.0]}
-                ),
-            ),
-            (
-                None,
-                "max",
-                pd.DataFrame(
-                    {"number_a": [4.0, 2.0, 3.0, 4.0], "number_b": [5.0, 6.0, 7.0, 7.0]}
-                ),
-            ),
-            (
-                None,
-                "min",
-                pd.DataFrame(
-                    {"number_a": [2.0, 2.0, 3.0, 4.0], "number_b": [5.0, 6.0, 7.0, 5.0]}
-                ),
-            ),
-        ],
-    )
-    def test_fill_na_imputes_numerical_na_correct(
-        self,
-        numerical_na: pd.DataFrame,
-        value: None,
-        strategy: str,
-        expected: pd.DataFrame,
-    ):
-        imputer = FillNA(value=value, strategy=strategy)
-        result = imputer.fit_transform(numerical_na)
-        pd.testing.assert_frame_equal(result, expected)
-
-    @pytest.mark.parametrize(
-        "value, strategy, expected",
-        [
-            (
-                "Unknown",
-                None,
-                pd.DataFrame(
-                    {
-                        "category_a": ["a1", "Unknown", "a3", "a1"],
-                        "category_b": ["Unknown", "b2", "b3", "b1"],
-                    },
-                    dtype="category",
-                ),
-            ),
-            (
-                None,
-                "most_freq",
-                pd.DataFrame(
-                    {
-                        "category_a": ["a1", "a1", "a3", "a1"],
-                        "category_b": ["b1", "b2", "b3", "b1"],
-                    },
-                    dtype="category",
-                ),
-            ),
-        ],
-    )
-    def test_fillna_imputes_pandas_categorical_correct(
-        self,
-        value: Any,
-        strategy: Any,
-        expected: pd.DataFrame,
-        categorical_na: pd.DataFrame,
-    ):
-        categorical_na["category_a"] = categorical_na["category_a"].astype("category")
-        categorical_na["category_b"] = categorical_na["category_b"].astype("category")
-
-        imputer = FillNA(value=value, strategy=strategy)
-        result = imputer.fit_transform(categorical_na)
-
-        pd.testing.assert_frame_equal(result, expected, check_categorical=False)
-
-    def test_fillna_works_cross_validated(self, train_iris_dataset):
-        model = create_model(FillNA(0))
-        result = model.score_estimator(train_iris_dataset, cv=2)
-        assert isinstance(result, Result)
-
-    def test_fillna_works_gridsearch(self, train_iris_dataset):
-        grid = create_gridsearch(FillNA(0))
-        model = Model(grid)
-        result = model.score_estimator(train_iris_dataset)
-        assert isinstance(result, Result)
-
-    def test_fillna_raises_when_imputing_numerically_on_strings(self):
-        df = pd.DataFrame(
-            {
-                "id": [1, 2, 3, 4],
-                "status": ["OK", "Error", "OK", "Error"],
-                "sales": [2000, 3000, 4000, np.nan],
-            }
-        )
-        fill_na = FillNA(strategy="mean")
-        with pytest.raises(
-            TransformerError,
-            match="column/columns have invalid types for strategy = mean",
-        ):
-            fill_na.fit_transform(df)
-
-        df["new_col"] = ["One", "Two", "Three", "Four"]
-
-        with pytest.raises(
-            TransformerError,
-            match="column/columns have invalid types for strategy = mean",
-        ):
-            fill_na.fit_transform(df)
-
-    def test_fillna_adds_is_na_column_when_imputing(self):
-        df = pd.DataFrame({"id": [1, 2, 3, 4], "sales": [2000, 3000, 4000, np.nan]})
-        fill_na = FillNA(strategy="mean", indicate_nan=True)
-        expected_cols = ["id", "sales", "sales_is_nan"]
-        result = fill_na.fit_transform(df)
-        assert result.columns.tolist() == expected_cols
-        assert np, all(result["sales_isna"] == [0, 0, 0, 1])
+    def test_works_without_args(self):
+        assert Select()
 
 
 class TestToCategorical:
+    def test_works_without_args(self):
+        assert ToCategorical()
+
     def test_to_categorical_returns_correct_dataframe(self, categorical: pd.DataFrame):
         to_cat = ToCategorical()
         result = to_cat.fit_transform(categorical)
@@ -318,6 +149,9 @@ class TestToCategorical:
 
 
 class TestBinner:
+    def test_works_without_args(self):
+        assert Binner()
+
     def test_binner_returns_correctly(self, numerical: pd.DataFrame):
         labels = ["1", "2", "3", "4"]
         binner = Binner(bins=[0, 1, 2, 3, 4], labels=labels)
@@ -354,6 +188,9 @@ class TestBinner:
 
 
 class TestRenamer:
+    def test_works_without_args(self):
+        assert Renamer()
+
     def test_renamer_returns_correctly(self, numerical: pd.DataFrame):
         new_col_names = ["test_a", "test_b"]
         renamer = Renamer(new_col_names)
@@ -393,6 +230,9 @@ class TestRenamer:
 
 
 class TestDateEncoder:
+    def test_works_without_args(self):
+        assert DateEncoder()
+
     def test_date_encoder_returns_correctly(self, dates: pd.DataFrame):
         date_coder = DateEncoder()
         result = date_coder.fit_transform(dates)
@@ -471,6 +311,9 @@ class TestDateEncoder:
 
 
 class TestFreqFeature:
+    def test_works_without_args(self):
+        assert FreqFeature()
+
     def test_freqfeature_returns_correctly(self, categorical: pd.DataFrame):
         freq_feature = FreqFeature()
         result = freq_feature.fit_transform(categorical)
@@ -545,6 +388,9 @@ class TestFeatureUnion:
 
 
 class TestStandardScaler:
+    def test_works_without_args(self):
+        assert DFStandardScaler()
+
     def test_can_reset_scaler_parameters(self):
         scaler = DFStandardScaler()
         scaler.scale_ = 0.7
@@ -615,6 +461,9 @@ class TestStandardScaler:
 
 
 class TestDFRowFunc:
+    def test_works_without_args(self):
+        assert DFRowFunc()
+
     @pytest.mark.parametrize(
         "strategy, match",
         [
@@ -707,6 +556,9 @@ def test_func_transformer_returns_correctly_numerical(
 
 
 class TestFuncTransformer:
+    def test_works_without_args(self):
+        assert FuncTransformer()
+
     def test_func_transformer_returns_correctly_on_categorical(
         self, categorical: pd.DataFrame
     ):
@@ -731,6 +583,9 @@ class TestFuncTransformer:
 
 
 class TestBinarize:
+    def test_works_without_args(self):
+        assert Binarize()
+
     def test_binarize_returns_correctly_on_categorical_na(
         self, categorical_na: pd.DataFrame
     ):
@@ -771,6 +626,9 @@ class TestRareFeatureEncoder:
     @pytest.fixture
     def categorical_int_and_string(self) -> pd.DataFrame:
         return pd.DataFrame({"categorical": [1, "a", "a", 2, "b", 1]})
+
+    def test_works_without_args(self):
+        assert RareFeatureEncoder()
 
     def test_rare_feature_encoder_that_transformed_data_and_input_data_same_shape(
         self, rare: RareFeatureEncoder, categorical_int_and_string: pd.DataFrame