+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+# v0.12.0
+- Permutation importance and Feature importance are now two different plotting methods.
+- `Model.test_estimators` now takes a `feature_pipeline` argument
+- Fixed a bug where `FillNA` did not create a `_is_na` column if the column didn't have a missing value
+- Implemented Bayesian Search for hyperparameter optimization
+- Added a `read_file` convenience method to `FileDataset` to read
+- Fixed a bug where `copy_to` failed between two instances of Sqlite based SQLDatasets
+- Fixed a bug where `ClassificationVisualize.confusion_matrix` would fail on multi-class problems due to wrong defaults
+- Added __repr__ to demodataset
+- Lift curve now can plot multi-class
+- Precision-Recall curve can now plot multi-class
+- ROC AUC curve can now plot multi-class
+- Fixed Binner to have a default value
+- Fixed FuncTransform to have a default value
+- `load_estimator` now uses default storage if nothing is passed
+- `Model.bayessearc` is now `Model.bayesiansearch`
+
 # v0.11.0
 - Added `load_demo_dataset` function
 - If the dataset has no train set `score_estimator` will now run `create_train_test` with default configurations
@@ -8,6 +25,7 @@
 - Can pass a feature pipeline to `Dataset.plot` methods, to apply preprocessing
 before visualization
 - New config implementation. If you need to reset the configuration, you should use `Model.config.reset_config()`
+
 
 # v0.10.3
 - Fixed typehints in Dataset